api_server/requirements.txt
amqp==5.1.1
anyascii==0.3.1
asgiref==3.6.0
astroid==2.7.3
audioread==3.0.0
Babel==2.12.1
billiard==3.6.4.0
black==23.1.0
bokeh==1.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
coqpit==0.0.17
coverage==7.2.1
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.7
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
fonttools==4.39.0
fsspec==2023.3.0
g2pK==0.9.4
g2pkk==0.1.2
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
idna==3.4
inflect==5.6.0
isort==5.12.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.1
jsonlines==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
konlpy==0.6.0
lazy-object-proxy==1.9.0
librosa==0.8.0
llvmlite==0.38.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.6.1
mecab-python3==1.0.5
mypy-extensions==1.0.0
mysqlclient==2.1.1
networkx==2.8.8
nltk==3.8.1
nose2==0.12.0
num2words==0.5.12
numba==0.55.2
numpy==1.22.4
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.1
pooch==1.7.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pylint==2.10.2
pynndescent==0.5.8
pyparsing==3.0.9
pypinyin==0.48.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-mecab-ko==1.3.3
python-mecab-ko-dic==2.1.1.post2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rabbitmq==0.2.0
regex==2022.10.31
requests==2.28.2
resampy==0.4.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
sqlparse==0.4.3
tensorboardX==2.6
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
tornado==6.2
tqdm==4.65.0
trainer==0.0.20
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
wrapt==1.12.1
